move controllers to folder, fix .env.sample
