+++ .env.sample
@@ -1,2 +1,2 @@
-REACT_APP_API_ENDPOINT=https://api.st-retrospect.dh-center.ru/graphql
+REACT_APP_API_ENDPOINT=https://api.st-retrospect.dh-center.ru/
 REACT_APP_MAPBOX_ACCESS_TOKEN=
